refactor: Validator 추가 하여 도메인에서 검증하는 로직 제거
