+++ src/main/java/racingcar/RacingCarApplication.java
@@ -1,6 +1,5 @@
 package racingcar;
 
-import racingcar.domain.Cars;
 import racingcar.domain.GameManager;
 import racingcar.domain.RandomNumberGenerator;
 import racingcar.view.InputView;
@@ -8,12 +7,12 @@ import racingcar.view.OutputView;
 
 public class RacingCarApplication {
 
-	public static void main(String[] args) {
-		GameManager gameManager = new GameManager(
-			new InputView(),
-			new OutputView(),
-			new RandomNumberGenerator());
+    public static void main(String[] args) {
+        GameManager gameManager = new GameManager(
+                new InputView(),
+                new OutputView(),
+                new RandomNumberGenerator());
 
-		gameManager.playGame();
-	}
+        gameManager.playGame();
+    }
 }
