+++ src/main/java/racingcar/domain/Cars.java
@@ -1,64 +1,58 @@
 package racingcar.domain;
 
-import static racingcar.exception.ErrorMessages.*;
+import racingcar.validator.CarValidator;
+import racingcar.validator.DefaultCarValidator;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Set;
+import java.util.*;
+
+import static racingcar.exception.ErrorMessages.CAR_NAME_DUPLICATE_EXCEPTION;
 
 public class Cars {
-	private final Set<Car> cars = new LinkedHashSet<>();
+    private final Set<Car> cars = new LinkedHashSet<>();
 
-	public void createCars(List<String> carNames, NumberGenerator numberGenerator) {
-		CarMovement carMovement = new CarMovement(numberGenerator);
-		validateCarsSizeOneOrZero(carNames);
-		for (String carName : carNames) {
-			generateCar(carName, carMovement);
-		}
-	}
+    public void createCars(List<String> carNames, NumberGenerator numberGenerator) {
+        CarValidator validator = new DefaultCarValidator();
+        validator.validateCarNamesSize(carNames);
 
-	public void moveCars() {
-		cars.forEach(Car::move);
-	}
+        CarMovement carMovement = new CarMovement(numberGenerator);
+        for (String carName : carNames) {
+            generateCar(carName, carMovement);
+        }
+    }
 
-	public List<String> findWinnerNames() {
-		List<String> winnerNames = new ArrayList<>();
-		int maxPosition = 0;
-		for (Car car : cars) {
-			if (car.getPosition() == maxPosition) {
-				winnerNames.add(car.getName());
-			}
-			if (car.getPosition() > maxPosition) {
-				maxPosition = car.getPosition();
-				winnerNames.clear();
-				winnerNames.add(car.getName());
-			}
-		}
-		return winnerNames;
-	}
+    public void moveCars() {
+        cars.forEach(Car::move);
+    }
 
-	public Set<Car> getCars() {
-		return Collections.unmodifiableSet(cars);
-	}
+    public List<String> findWinnerNames() {
+        List<String> winnerNames = new ArrayList<>();
+        int maxPosition = 0;
+        for (Car car : cars) {
+            if (car.getPosition() == maxPosition) {
+                winnerNames.add(car.getName());
+            }
+            if (car.getPosition() > maxPosition) {
+                maxPosition = car.getPosition();
+                winnerNames.clear();
+                winnerNames.add(car.getName());
+            }
+        }
+        return winnerNames;
+    }
 
-	private void generateCar(String name, CarMovement carMovement) {
-		Car newCar = new Car(name, carMovement);
-		validateDuplicatedCarName(newCar);
-		cars.add(newCar);
-	}
+    public Set<Car> getCars() {
+        return Collections.unmodifiableSet(cars);
+    }
 
-	private void validateCarsSizeOneOrZero(List<String> carNames) {
-		int carsSize = carNames.size();
-		if (carsSize == 0 || carsSize == 1) {
-			throw new IllegalArgumentException(CARS_SIZE_EXCEPTION.getMessage());
-		}
-	}
+    private void generateCar(String name, CarMovement carMovement) {
+        Car newCar = new Car(name, carMovement);
+        validateDuplicatedCarName(newCar);
+        cars.add(newCar);
+    }
 
-	private void validateDuplicatedCarName(Car newCar) {
-		if (cars.contains(newCar)) {
-			throw new IllegalArgumentException(CAR_NAME_DUPLICATE_EXCEPTION.getMessage());
-		}
-	}
+    private void validateDuplicatedCarName(Car newCar) {
+        if (cars.contains(newCar)) {
+            throw new IllegalArgumentException(CAR_NAME_DUPLICATE_EXCEPTION.getMessage());
+        }
+    }
 }
