+++ src/main/java/racingcar/domain/GameRound.java
@@ -1,19 +1,19 @@
 package racingcar.domain;
 
 public class GameRound {
-	private final int totalRound;
-	private int currRound;
+    private final int totalRound;
+    private int currRound;
 
-	public GameRound(int totalRound) {
-		this.totalRound = totalRound;
-		this.currRound = 0;
-	}
+    public GameRound(int totalRound) {
+        this.totalRound = totalRound;
+        this.currRound = 0;
+    }
 
-	public void increaseRound() {
-		this.currRound++;
-	}
+    public void increaseRound() {
+        this.currRound++;
+    }
 
-	public boolean isEnd() {
-		return this.currRound == this.totalRound;
-	}
+    public boolean isEnd() {
+        return this.currRound == this.totalRound;
+    }
 }
