+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -3,9 +3,9 @@ package racingcar.domain;
 import java.util.Random;
 
 public class RandomNumberGenerator implements NumberGenerator {
-	private static final Random RANDOM = new Random();
+    private static final Random RANDOM = new Random();
 
-	public int generateNumber() {
-		return RANDOM.nextInt(10);
-	}
+    public int generateNumber() {
+        return RANDOM.nextInt(10);
+    }
 }
