+++ src/main/java/racingcar/dto/CarDTO.java
@@ -3,19 +3,19 @@ package racingcar.dto;
 import racingcar.domain.Car;
 
 public class CarDTO {
-	private final String carName;
-	private final int carPosition;
+    private final String carName;
+    private final int carPosition;
 
-	public CarDTO(Car car) {
-		this.carName = car.getName();
-		this.carPosition = car.getPosition();
-	}
+    public CarDTO(Car car) {
+        this.carName = car.getName();
+        this.carPosition = car.getPosition();
+    }
 
-	public String getCarName() {
-		return carName;
-	}
+    public String getCarName() {
+        return carName;
+    }
 
-	public int getCarPosition() {
-		return carPosition;
-	}
+    public int getCarPosition() {
+        return carPosition;
+    }
 }
