+++ src/main/java/racingcar/dto/GameResultResponse.java
@@ -4,13 +4,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class GameResultResponse {
-	private final List<String> winnerNames = new ArrayList<>();
+    private final List<String> winnerNames = new ArrayList<>();
 
-	public GameResultResponse(List<String> winnerNames) {
-		this.winnerNames.addAll(winnerNames);
-	}
+    public GameResultResponse(List<String> winnerNames) {
+        this.winnerNames.addAll(winnerNames);
+    }
 
-	public List<String> getWinnerNames() {
-		return winnerNames;
-	}
+    public List<String> getWinnerNames() {
+        return winnerNames;
+    }
 }
