+++ src/main/java/racingcar/dto/RoundResultResponse.java
@@ -1,21 +1,21 @@
 package racingcar.dto;
 
+import racingcar.domain.Car;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-import racingcar.domain.Car;
-
 public class RoundResultResponse {
-	private final List<CarDTO> roundResult = new ArrayList<>();
+    private final List<CarDTO> roundResult = new ArrayList<>();
 
-	public RoundResultResponse(Set<Car> cars) {
-		for (Car car : cars) {
-			roundResult.add(new CarDTO(car));
-		}
-	}
+    public RoundResultResponse(Set<Car> cars) {
+        for (Car car : cars) {
+            roundResult.add(new CarDTO(car));
+        }
+    }
 
-	public List<CarDTO> getRoundResult() {
-		return roundResult;
-	}
+    public List<CarDTO> getRoundResult() {
+        return roundResult;
+    }
 }
