+++ src/main/java/racingcar/validator/CarValidator.java
@@ -1,0 +1,9 @@
+package racingcar.validator;
+
+import java.util.List;
+
+public interface CarValidator {
+    void validateCarName(String name);
+
+    void validateCarNamesSize(List<String> carNames);
+}
