+++ src/main/java/racingcar/validator/GameRoundValidator.java
@@ -1,0 +1,5 @@
+package racingcar.validator;
+
+public interface GameRoundValidator {
+    void validateGameRoundInput(String inputGameRound);
+}
