+++ src/test/java/racingcar/domain/CarMovementTest.java
@@ -1,15 +1,15 @@
 package racingcar.domain;
 
-import static org.assertj.core.api.Assertions.*;
-import static racingcar.config.CarMovementConfig.*;
-
-import java.util.stream.Stream;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+
+import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static racingcar.config.CarMovementConfig.*;
 
 class CarMovementTest {
 	NumberGenerator numberGenerator;
