+++ src/test/java/racingcar/utils/RandomNumberGeneratorTest.java
@@ -1,12 +1,11 @@
 package racingcar.utils;
 
-import static org.assertj.core.api.Assertions.*;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
-
 import racingcar.domain.NumberGenerator;
 import racingcar.domain.RandomNumberGenerator;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 class RandomNumberGeneratorTest {
 	final NumberGenerator numberGenerator = new RandomNumberGenerator();
